[468] Ядро / Разное: Исправлено сбрасывание талантов при логине Исправлен инспект достижений Исправлено предложение воскрешения Печатаем StackTrace когда пытаемся отправить NULL\UNKNOWN опкод

+++ src/server/game/Handlers/QueryHandler.cpp
@@ -397,9 +397,9 @@ void WorldSession::HandleCorpseMapPositionQuery(WorldPacket& recvData)
     recvData >> unk;
 
     WorldPacket data(SMSG_CORPSE_MAP_POSITION_QUERY_RESPONSE, 4+4+4+4);
-    data << float(0);
-    data << float(0);
-    data << float(0);
+    data << int(0);
+    data << int(0);
+    data << int(0);
     data << float(0);
     SendPacket(&data);
 }
